More wiring, did away with Keyed/Singleton dichotomy.
It wasn't adding a lot of value, and you can just create one instance of an
entity and call it a singleton if that's how you like to roll.

This removes the "auto-create" nature of singletons, but I think that was a bad
idea in the first place. You often need to init entities, so just create your
singleton during system init along with everything else:

if (keys(SomeSingleton).isEmpty()) create(SomeSingleton) { single ->
  // init single...
}