We need SourceDB to be a superset of DDB API.
Otherwise life becomes a PITA for clients. So we'll just reject a request to
obtain a bare service on the server. We'll eventually add another way to get a
service on the server where you supply an execution context into which to
deliver results, so that we can maintain thread sanity.