A whole lotta wiring up.
Services now have marshallers and dispatchers generated for them, and we have
better support for serializing arbitrary graphs of maps of maps of lists of
maps of whatever. However, that latter business is going to require further
work because right now we trigger off an exact match of the concrete collection
type which won't cut it. The current tests are already failing for lack of a
serializer for Arrays.ArrayList. Blah.