Revamp prop change apply.
Now we know when we're applying a prop change that has come in from over the
network versus a prop change that was initiated locally.

This allows us to reinstate the freakoutery if you try to change an entity on
the client. We're about to stop freaking out about that (or at least enable
some changes to be made from the client), but this is a good stepping stone.