Wire up client-initiated property change.
This includes a rudimentary new ACL mechanism which rejects all
client-initiated prop changes by default, but can be configured per-entity type
to allow certain changes. Right now it's on an entity by entity basis rather
than a property by property basis, but we'll probably want to eventually enable
the latter.

We'll also undoubtedly eventually want DDB-subscribe ACLs so that clients can't
just subscribe to any DDB they want, and we may even want "filters" on the
information in a particular DDB such that clients don't see everything in a
DDB. This will be trickier and I'm not yet sure whether I want to go there or
to instead say, "put private stuff in a separate DDB".

The latter is easier for me the DDB implementor, but probably a PITA for the
DDB user. However, I don't want to open a pandora's box of complexity and
unexpected sharp edges by tossing out the simple "a DDB is a collection of
key/value stores that is mirrored from server to client as changes occur"
mental model.

Anyway...